Improves harvesting
This patches the harvesting for SOS services. I removed the compliance
checker scores for SOS, they don't mean much right now. I hopefully
improved performance by caching the outputFormat so that if the
harvester finds an outputFormat for a particular service, it should keep
using that one instead of trying the list of them.

The formatting of the templates is changed so that there's more colors
than just red and clear. If a service is partially available (some
procedures worked) it will be yellow. If the service was unavailable it
will be gray. If it failed to parse correctly it's red.